# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantitative Trading Research Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type StrategyFileType =
  (typeof StrategyFileType)[keyof typeof StrategyFileType];

export const StrategyFileType = {
  mq5: "mq5",
  pine: "pine",
  python: "python",
  mql4: "mql4",
  txt: "txt",
  other: "other",
} as const;

export interface Strategy {
  id: number;
  name: string;
  description?: string;
  fileType: StrategyFileType;
  fileName: string;
  fileContent: string;
  strategyType?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AnalysisOptions {
  markets?: string[];
  timeframes?: string[];
  yearsOfData?: number;
  trainSplit?: number;
  validationSplit?: number;
}

export type AnalysisRunStatus =
  (typeof AnalysisRunStatus)[keyof typeof AnalysisRunStatus];

export const AnalysisRunStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface AnalysisRun {
  id: number;
  strategyId: number;
  status: AnalysisRunStatus;
  currentStep?: number | null;
  totalSteps: number;
  options: AnalysisOptions;
  summary?: string | null;
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type StepResultStatus =
  (typeof StepResultStatus)[keyof typeof StepResultStatus];

export const StepResultStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export type StepResultMetrics = { [key: string]: unknown } | null;

export interface StepResult {
  id: number;
  runId: number;
  stepNumber: number;
  stepName: string;
  status: StepResultStatus;
  findings?: string | null;
  metrics?: StepResultMetrics;
  recommendations?: string | null;
  startedAt?: string | null;
  completedAt?: string | null;
}

export type UploadStrategyBody = {
  file: Blob;
  name?: string;
  description?: string;
};