# lib/api-zod/src/generated/types/analysisOptions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantitative Trading Research Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisOptions {
  markets?: string[];
  timeframes?: string[];
  yearsOfData?: number;
  trainSplit?: number;
  validationSplit?: number;
}