# lib/api-zod/src/generated/types/analysisRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantitative Trading Research Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisOptions } from "./analysisOptions";
import type { AnalysisRunStatus } from "./analysisRunStatus";

export interface AnalysisRun {
  id: number;
  strategyId: number;
  status: AnalysisRunStatus;
  currentStep?: number | null;
  totalSteps: number;
  options: AnalysisOptions;
  summary?: string | null;
  errorMessage?: string | null;
  createdAt: Date;
  updatedAt: Date;
}