# lib/api-zod/src/generated/types/stepResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantitative Trading Research Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StepResultMetrics } from "./stepResultMetrics";
import type { StepResultStatus } from "./stepResultStatus";

export interface StepResult {
  id: number;
  runId: number;
  stepNumber: number;
  stepName: string;
  status: StepResultStatus;
  findings?: string | null;
  metrics?: StepResultMetrics;
  recommendations?: string | null;
  startedAt?: Date | null;
  completedAt?: Date | null;
}