# lib/api-zod/src/generated/types/strategy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quantitative Trading Research Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StrategyFileType } from "./strategyFileType";

export interface Strategy {
  id: number;
  name: string;
  description?: string;
  fileType: StrategyFileType;
  fileName: string;
  fileContent: string;
  strategyType?: string | null;
  createdAt: Date;
  updatedAt: Date;
}